Refactor changes
Renaming bots ✓
Adding packages ✓
