--- Erdogan.java
@@ -1,4 +1,0 @@
-package Dictators;
-
-public class Erdogan extends Dictator {
-}
