+++ KimJongUn.java
@@ -1,4 +1,6 @@
 package Dictators;
 
+import Dictators.Model.Dictator;
+
 public class KimJongUn extends Dictator {
 }
