+++ Putin.java
@@ -2,5 +2,5 @@ package Dictators;
 
 import Dictators.Model.Dictator;
 
-public class Trump extends Dictator {
+public class Putin extends Dictator {
 }
